+++ src/New/Type/AttachEffectTypeClass.h
@@ -43,6 +43,9 @@ public:
 	ValueableVector<WeaponTypeClass*> WeaponRange_DisallowWeapons;
 	Valueable<double> Crit_Multiplier;
 	Valueable<double> Crit_ExtraChance;
+	Valueable<double> Crit_ExtraDamage_Multiplier;
+	Valueable<bool> Crit_ExtraDamage_AllowNegative;
+	Valueable<int> Crit_ExtraDamage_Bonus;
 	ValueableVector<WarheadTypeClass*> Crit_AllowWarheads;
 	ValueableVector<WarheadTypeClass*> Crit_DisallowWarheads;
 	Nullable<WeaponTypeClass*> RevengeWeapon;
@@ -83,6 +86,9 @@ public:
 		, WeaponRange_DisallowWeapons {}
 		, Crit_Multiplier { 1.0 }
 		, Crit_ExtraChance { 0.0 }
+		, Crit_ExtraDamage_Multiplier { 1.0 }
+		, Crit_ExtraDamage_Bonus { 0 }
+		, Crit_ExtraDamage_AllowNegative { true }
 		, Crit_AllowWarheads {}
 		, Crit_DisallowWarheads {}
 		, RevengeWeapon {}